backend/requirements.txt
Django==3.2.20
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.12.4
djoser==2.1.0
flake8==6.0.0
fpdf2==2.7.4
gunicorn==20.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
PyYAML==6.0
